# FIX: support_enumeration: Use `_numba_linalg_solve`