--- a/docs/source/util.rst
+++ b/docs/source/util.rst
@@ -7,5 +7,6 @@ Utilities
    util/array
    util/common_messages
    util/notebooks
+   util/numba
    util/random
    util/timing