--- a/docs/source/util/numba.rst
+++ b/docs/source/util/numba.rst
@@ -1,0 +1,7 @@
+numba
+=====
+
+.. automodule:: quantecon.util.numba
+    :members:
+    :undoc-members:
+    :show-inheritance: