--- a/quantecon/game_theory/support_enumeration.py
+++ b/quantecon/game_theory/support_enumeration.py
@@ -12,11 +12,8 @@ Tardos, and V. Vazirani eds., Algorithmic Game Theory, 2007.
 
 """
 import numpy as np
-import numba
 from numba import jit
-
-
-EPS = np.finfo(float).eps
+from ..util.numba import _numba_linalg_solve
 
 
 def support_enumeration(g):
@@ -93,32 +90,28 @@ def _support_enumeration_gen(payoff_matrix0, payoff_matrix1):
 
     for k in range(1, n_min+1):
         supps = (np.arange(k), np.empty(k, np.int_))
-        actions = (np.empty(k), np.empty(k))
+        actions = (np.empty(k+1), np.empty(k+1))
         A = np.empty((k+1, k+1))
-        A[:-1, -1] = -1
-        A[-1, :-1] = 1
-        A[-1, -1] = 0
-        b = np.zeros(k+1)
-        b[-1] = 1
+
         while supps[0][-1] < nums_actions[0]:
             supps[1][:] = np.arange(k)
             while supps[1][-1] < nums_actions[1]:
                 if _indiff_mixed_action(payoff_matrix0, supps[0], supps[1],
-                                        A, b, actions[1]):
+                                        A, actions[1]):
                     if _indiff_mixed_action(payoff_matrix1, supps[1], supps[0],
-                                            A, b, actions[0]):
+                                            A, actions[0]):
                         out = (np.zeros(nums_actions[0]),
                                np.zeros(nums_actions[1]))
                         for p, (supp, action) in enumerate(zip(supps,
                                                                actions)):
-                            out[p][supp] = action
+                            out[p][supp] = action[:-1]
                         yield out
                 _next_k_array(supps[1])
             _next_k_array(supps[0])
 
 
 @jit(nopython=True)
-def _indiff_mixed_action(payoff_matrix, own_supp, opp_supp, A, b, out):
+def _indiff_mixed_action(payoff_matrix, own_supp, opp_supp, A, out):
     """
     Given a player's payoff matrix `payoff_matrix`, an array `own_supp`
     of this player's actions, and an array `opp_supp` of the opponent's
@@ -127,8 +120,7 @@ def _indiff_mixed_action(payoff_matrix, own_supp, opp_supp, A, b, out):
     among the actions in `own_supp`, if any such exists. Return `True`
     if such a mixed action exists and actions in `own_supp` are indeed
     best responses to it, in which case the outcome is stored in `out`;
-    `False` otherwise. Arrays `A` and `b` are used in intermediate
-    steps.
+    `False` otherwise. Array `A` is used in intermediate steps.
 
     Parameters
     ----------
@@ -142,17 +134,11 @@ def _indiff_mixed_action(payoff_matrix, own_supp, opp_supp, A, b, out):
         Array containing the opponent's action indices, of length k.
 
     A : ndarray(float, ndim=2)
-        Array used in intermediate steps, of shape (k+1, k+1). The
-        following values must be assigned in advance: `A[:-1, -1] = -1`,
-        `A[-1, :-1] = 1`, and `A[-1, -1] = 0`.
-
-    b : ndarray(float, ndim=1)
-        Array used in intermediate steps, of shape (k+1,). The following
-        values must be assigned in advance `b[:-1] = 0` and `b[-1] = 1`.
+        Array used in intermediate steps, of shape (k+1, k+1).
 
     out : ndarray(float, ndim=1)
-        Array of length k to store the k nonzero values of the desired
-        mixed action.
+        Array of length k+1 to store the k nonzero values of the desired
+        mixed action in `out[:-1]` (and the payoff value in `out[-1]`.)
 
     Returns
     -------
@@ -163,15 +149,21 @@ def _indiff_mixed_action(payoff_matrix, own_supp, opp_supp, A, b, out):
     m = payoff_matrix.shape[0]
     k = len(own_supp)
 
-    A[:-1, :-1] = payoff_matrix[own_supp, :][:, opp_supp]
-    if _is_singular(A):
-        return False
+    for i in range(k):
+        for j in range(k):
+            A[j, i] = payoff_matrix[own_supp[i], opp_supp[j]]  # transpose
+    A[:-1, -1] = 1
+    A[-1, :-1] = -1
+    A[-1, -1] = 0
+    out[:-1] = 0
+    out[-1] = 1
 
-    sol = np.linalg.solve(A, b)
-    if (sol[:-1] <= 0).any():
+    r = _numba_linalg_solve(A, out)
+    if r != 0:  # A: singular
         return False
-    out[:] = sol[:-1]
-    val = sol[-1]
+    if (out[:-1] <= 0).any():
+        return False
+    val = out[-1]
 
     if k == m:
         return True
@@ -268,27 +260,3 @@ def _next_k_array(a):
         pos += 1
 
     return a
-
-
-@jit(nopython=True, cache=True)
-def _is_singular(a):
-    """
-    Determine whether matrix `a` is numerically singular, by checking
-    its singular values.
-
-    Parameters
-    ----------
-    a : ndarray(float, ndim=2)
-        2-dimensional array of floats.
-
-    Returns
-    -------
-    bool
-        Whether `a` is numerically singular.
-
-    """
-    s = numba.targets.linalg._compute_singular_values(a)
-    if s[-1] <= s[0] * EPS:
-        return True
-    else:
-        return False