* It turned out that the workaround in item 3 in https://github.com/QuantEcon/QuantEcon.py/pull/263#issue-178525924 did not work properly in all cases.
* This fix replaces "SVD + linalg.solve" with `_numba_linalg_solve`, which calls directly Numba internal `numba_xgesv`.
* As a bonus, this leads to about 9x speedup.

```py
ns = [8, 9, 10, 11]
seed = 1234
for n in ns:
    g = gt.random_game((n, n), random_state=seed)
    %timeit gt.support_enumeration(g)
```

* Current version:
  * n = 8: `99.1 ms`
  * n = 9: `391 ms`
  * n = 10: `1.58 s`
  * n = 11: `6.56 s`

* This PR:
  * n = 8: `11.6 ms`
  * n = 9: `43 ms`
  * n = 10: `179 ms`
  * n = 11: `727 ms`

```py
numba.__version__
'0.30.1'
```