@mmcky 2017-05-29
Thanks @oyamad. I just have one ``inline`` question/comment.